Started to implement cgfx format (very messy atm...), also some tweaks on the deserializar to support cgfx too (which hopefully didn't broke h3d), fixed a issue with quantization on vertex, fixed issues parsing SMDs and OBJs containing names with spaces, other small fixes

+++ SPICA/PICA/Commands/PICAAttributeName.cs
@@ -1,6 +1,6 @@
 ﻿namespace SPICA.PICA.Commands
 {
-    public enum PICAAttributeName : byte
+    public enum PICAAttributeName : uint
     {
         Position,
         Normal,
